Use native Hugo templating to generate crosswalks
Instead of .Rmd files, which require having .html files as input to Hugo,
which need to be committed to git.

This makes it easier to keep the website up-to-date with the main
crosswalk.

This also updates `.github/workflows/deploy.yml` to run daily, in order
to apply such changes from github.com/codemeta/codemeta without the need for
specific commits in this repository (github.com/codemeta/codemeta.github.io/)